adding minor refactoring changes
