+++ compiler/src/parser/resolver.rs
@@ -24,7 +24,8 @@ use crate::scope::concrete::ConcreteTypesTuple;
 use crate::scope::core::{LookupData, LookupResult};
 use crate::scope::errors::GenericTypeArgsCheckError;
 use crate::scope::mangled::MangledIdentifierName;
-use crate::scope::semantic_db::{ConcreteSymbolDataEntry, SemanticStateDatabase, SymbolDataEntry};
+use crate::scope::semantic_db::SemanticStateDatabase;
+use crate::scope::symbol::core::{ConcreteSymbolDataEntry, SymbolDataEntry};
 use crate::scope::symbol::function::FunctionSymbolData;
 use crate::scope::symbol::function::{CallableKind, CallablePrototypeData};
 use crate::scope::symbol::interfaces::InterfaceSymbolData;
